requirements.txt
pydantic
pyyaml
fastapi
uvicorn